rendering card with search feature in new page
